refactor: extract HalkyonEntity from delete to reuse it in other commands
